+++ src/types/associations/shared.ts
@@ -216,12 +216,14 @@ export type ColumnNamesAccountingForJoinedAssociations<
   RootTableAlias extends string,
 > = JoinedAssociations['length'] extends 0
   ? TableColumnNames<DB, RootTableName> // no associations; simply return the un-namespaced columns for the root table
-  : JoinedAssociationColumnNames<
-      JoinedAssociations,
-      DB,
-      // namespace columns for the root table
-      AssociationNameToDotReferencedColumns<DB, RootTableName & keyof DB, RootTableAlias>
-    >
+  :
+      | TableColumnNames<DB, RootTableName>
+      | JoinedAssociationColumnNames<
+          JoinedAssociations,
+          DB,
+          // namespace columns for the root table
+          AssociationNameToDotReferencedColumns<DB, RootTableName & keyof DB, RootTableAlias>
+        >
 type JoinedAssociationColumnNames<
   JoinedAssociations extends Readonly<JoinedAssociation[]>,
   DB,